+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Unreleased (0.3.0.0)
 
-* Optimized rendering of errors ([#5](https://github.com/1Computer1/errata/pull/5)). Huge thanks to [RiugaBachi](https://github.com/RiugaBachi) for this!
+* Optimized rendering of errors ([#5](https://github.com/1Computer1/errata/pull/5)). Huge thanks to [RiugaBachi](https://github.com/RiugaBachi) for this! This also adds a `Monoid` constraint to `Source`.
 
 * Reworked the pretty printer so that it no longer prints trailing whitespace in most places.